feat: Added fallback ability to resolve bib-entries from doi.org.
One can retrieve bib-entries from doi.org with the following HTTP
request:
$ curl -sLH "Accept: application/x-bibtex; charset=utf-8" https://doi.org/10.1007/11880561_13

We usually prefer to fetch bib-entries from dblp.org as they are better
maintained. However, if retrieval from dblp.org fails (e.g., because the
server is down or because the entry is not available there), we can now
fallback to the bib-entry offered at doi.org. When --prefer-doi-org is
used in the command line, doi.org is asked first.